Lesson16 HW16

+++ src/ru/javawebinar/basejava/web/ResumeServlet.java
@@ -2,7 +2,11 @@ package ru.javawebinar.basejava.web;
 
 import ru.javawebinar.basejava.Config;
 import ru.javawebinar.basejava.model.ContactType;
+import ru.javawebinar.basejava.model.ListSection;
 import ru.javawebinar.basejava.model.Resume;
+import ru.javawebinar.basejava.model.Section;
+import ru.javawebinar.basejava.model.SectionType;
+import ru.javawebinar.basejava.model.TextSection;
 import ru.javawebinar.basejava.storage.Storage;
 
 import javax.servlet.ServletConfig;
@@ -23,10 +27,23 @@ public class ResumeServlet extends HttpServlet {
 
     @Override
     protected void doPost(HttpServletRequest request, HttpServletResponse response) throws javax.servlet.ServletException, IOException {
+        // Совмещаем здесь Create и Update
         request.setCharacterEncoding("UTF-8");
         String uuid = request.getParameter("uuid");
         String fullName = request.getParameter("fullName");
-        Resume r = storage.get(uuid);
+        if (uuid == null
+                || uuid.trim().isEmpty()
+                || fullName == null
+                || fullName.trim().isEmpty()) {
+            response.sendRedirect("resume");
+            return;
+        }
+        uuid = uuid.trim();
+        fullName = fullName.trim();
+
+        boolean isNew = "true".equals(request.getParameter("isNew"));
+        Resume r = isNew ? new Resume("") : storage.get(uuid);
+
         r.setFullName(fullName);
         for (ContactType type : ContactType.values()) {
             String value = request.getParameter(type.name());
@@ -36,7 +53,23 @@ public class ResumeServlet extends HttpServlet {
                 r.getContacts().remove(type);
             }
         }
-        storage.update(r);
+        for (SectionType type : SectionType.values()) {
+            String value = request.getParameter(type.name());
+            Section section = mapToSection(type, value);
+            if (section != null) {
+                r.addSection(type, section);
+            } else {
+                // т.к. пока не реализовал передачу орг секций из edit в сервлет,
+                // то при обновлении резюме они удаляются :-(
+                r.getSections().remove(type);
+            }
+        }
+        if (isNew) {
+            storage.save(r);
+        } else {
+            storage.update(r);
+        }
+
         response.sendRedirect("resume");
     }
 
@@ -44,28 +77,59 @@ public class ResumeServlet extends HttpServlet {
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws javax.servlet.ServletException, IOException {
         String uuid = request.getParameter("uuid");
         String action = request.getParameter("action");
+
         if (action == null) {
             request.setAttribute("resumes", storage.getAllSorted());
             request.getRequestDispatcher("/WEB-INF/jsp/list.jsp").forward(request, response);
             return;
         }
-        Resume r;
+
         switch (action) {
             case "delete":
-                storage.delete(uuid);
+                if (uuid != null) storage.delete(uuid);
                 response.sendRedirect("resume");
                 return;
             case "view":
+                if (uuid != null) {
+                    request.setAttribute("resume", storage.get(uuid));
+                    request.getRequestDispatcher("/WEB-INF/jsp/view.jsp").forward(request, response);
+                } else {
+                    response.sendRedirect("resume");
+                }
+                return;
             case "edit":
-                r = storage.get(uuid);
-                break;
+                boolean isNew = uuid == null;
+                Resume r = isNew ? new Resume("") : storage.get(uuid);
+
+                request.setAttribute("resume", r);
+                request.setAttribute("isNew", isNew);
+                request.getRequestDispatcher("/WEB-INF/jsp/edit.jsp").forward(request, response);
+                return;
             default:
                 throw new IllegalArgumentException("Action " + action + " is illegal");
         }
-        request.setAttribute("resume", r);
-        request.getRequestDispatcher(
-                ("view".equals(action) ? "/WEB-INF/jsp/view.jsp" : "/WEB-INF/jsp/edit.jsp")
-        ).forward(request, response);
+
+    }
+
+    private Section mapToSection(SectionType type, String value) {
+        if (value == null || value.isEmpty()) return null;
+
+        switch (type) {
+            case PERSONAL:
+            case OBJECTIVE:
+                // TextSection
+                return new TextSection(value);
+            case ACHIEVEMENT:
+            case QUALIFICATIONS:
+                // ListSection
+                return value.isEmpty() ? new ListSection() : new ListSection(value.split("\n"));
+            case EXPERIENCE:
+            case EDUCATION:
+                // OrganizationSection
+                return null;
+            default:
+                throw new UnsupportedOperationException("Unsupported section type " + type.name());
+        }
     }
 
 }
